requirements.txt
openai-agents
ollama
python-dotenv
requests
pydantic